Fix patient creation flow and UI links
Update patient create/edit flow, UI links, and build classpath.

Key changes:
- header.jsp: update technician link to /lab-queue and patient menu link to /patient-health-dashboard.
- nbproject/project.properties: change jakartaee-migration jar reference to 1.0.7.
- PatientPortalDAO: change addPatient return type from int to long and return generated key as long.
- PatientsServlet: refactor parameter parsing and validation (parse userID before constructing Patient, handle patientID safely for edit vs create), preserve patient and error attributes on validation failure so the form is re-displayed with user input, and redirect receptionist to appointment flow with the new patient id. Also adjust role checks to use the existing role representation.
- createPatients.jsp: cleanup markup, use patient.patientId for hidden field, fix input/value bindings and error messages, update gender selection to consider existing patient or submitted value, and simplify role-based submit buttons (receptionist sees "Chọn ca", others see create/edit appropriately).

Note: addPatient signature change is breaking for callers; servlet updated accordingly.

+++ src/java/dal/PatientPortalDAO.java
@@ -397,7 +397,7 @@ public class PatientPortalDAO extends DBContext {
         }
     }
 
-    public int addPatient(Patient patient) {
+    public long addPatient(Patient patient) {
 
         String sql = "INSERT INTO patients (user_id, full_name, phone, dob, email, gender) VALUES (?, ?, ?, ?, ?, ?)";
 
@@ -427,7 +427,7 @@ public class PatientPortalDAO extends DBContext {
 
             ResultSet rs = ps.getGeneratedKeys();
             if (rs.next()) {
-                return rs.getInt(1);
+                return rs.getLong(1);
             }
 
         } catch (Exception e) {
